requirements.txt
torchvision>=0.6.1
